src/gt4sd/frameworks/molgx/requirements.txt
# pypi requirements
matplotlib>=3.2.0
pandas>=1.1.0,<1.4
scikit-learn>=0.23.2,<=0.24.2
ipython>=7.22.0,<8.0.0
rdkit-pypi>=2020.9.5.2,<=2021.9.4
scipy>=1.6.0,<1.8.0
numpy>=1.16.5,<=1.20.2
